v7.0.5

See merge request scotch/scotch!21

+++ src/libscotch/dgraph_halo.c
@@ -1,4 +1,4 @@
-/* Copyright 2007-2009,2011,2014,2020,2021,2023 IPB, Universite de Bordeaux, INRIA & CNRS
+/* Copyright 2007-2009,2011,2014,2020,2021,2023,2024 IPB, Universite de Bordeaux, INRIA & CNRS
 **
 ** This file is part of the Scotch software package for static mapping,
 ** graph partitioning and sparse matrix ordering.
@@ -56,7 +56,7 @@
 /**                # Version 6.1  : from : 05 apr 2021     **/
 /**                                 to   : 18 dec 2021     **/
 /**                # Version 7.0  : from : 03 sep 2020     **/
-/**                                 to   : 17 jan 2023     **/
+/**                                 to   : 09 aug 2024     **/
 /**                                                        **/
 /************************************************************/
 
@@ -227,7 +227,7 @@ MPI_Request ** const          requptr)            /* Pointer to local request ar
     return (1);
   }
 
-  dgraphHaloFill (grafptr, attrgsttab, attrglbsiz, *attrsndptr, *senddspptr, grafptr->procsndtab); /* Fill data arrays */
+  dgraphHaloFill (grafptr, attrgsttab, (int) attrglbsiz, *attrsndptr, *senddspptr, grafptr->procsndtab); /* Fill data arrays */
 
   recvdsptab = *recvdspptr;
   procrcvtab = grafptr->procrcvtab;
