+++ src/libscotch/library_error.c
@@ -1,4 +1,4 @@
-/* Copyright 2004,2007,2008,2011,2023 IPB, Universite de Bordeaux, INRIA & CNRS
+/* Copyright 2004,2007,2008,2011,2023,2024 IPB, Universite de Bordeaux, INRIA & CNRS
 **
 ** This file is part of the Scotch software package for static mapping,
 ** graph partitioning and sparse matrix ordering.
@@ -48,7 +48,7 @@
 /**                # Version 5.1  : from : 27 sep 2008     **/
 /**                                 to   : 17 jul 2011     **/
 /**                # Version 7.0  : from : 21 jan 2023     **/
-/**                                 to   : 21 jan 2023     **/
+/**                                 to   : 09 aug 2024     **/
 /**                                                        **/
 /************************************************************/
 
@@ -78,7 +78,7 @@ void
 SCOTCH_errorProg (
 const char * const          progstr)              /*+ Program name +*/
 {
-  int                 charnbr;
+  size_t              charnbr;
   const char *        nsrcptr;
   char *              ndstptr;
